imported assets, minor changes to betting, pot prefab created;

+++ SW_PDX_GAME/Online_game/Assets/Scripts/MultiplayerGame.cs
@@ -191,7 +191,7 @@ namespace GoFish
         }
 
         //****************** User Interaction *********************//
-        public override void OnBetSelected(float sliderVal)
+        public override void OnBetSelected(double sliderVal)
         {
             if (gameState == GameState.Bet && localPlayer == currentTurnPlayer)
             {
