+++ SW_PDX_GAME/Online_game/Assets/Scripts/Player.cs
@@ -19,7 +19,7 @@ namespace GoFish
         public int BuyIns = 1;
         //public int TournamentAddOn = 0;
         //public float CashAddOn = 0;
-        public float StackAmt;
+        public double StackAmt;
         public double bet;
         int numberOfDisplayingCards;
         int numberOfBooks;
